Dev snapshot.
Here I've decided the projection MUST have a Datum and this carries
all the ellipsoid parameters. The point may or may not have a datum.
If it does not, then it will inherit the datum from the projection.
If the point does have a datum and it is not the same as the
projection datum then the point will be datum shifted before any
transforms are done.